Update fleet_size_estimation.m
I have added my part of the code. I am not sure about pull request. So, I am just pasting it in your branch.
Its not very elegant but it works.

Now we have the number of round trips and minimum empty trips with the assumption that the capacity at B is just sufficient to carry the passengers from B to A. Also, the number of unused cars every hour is also calculated. This does not account for cars unused when the trip is completed. Rather, the cars unused for the whole hour.